refactor: prayerRoom Dir 구조 변경 #25
- rooms로 통일
- invitation 도메인 생성
