+++ app/prayers/[id]/_components/index/PrayerRoomBody.tsx
@@ -1,0 +1,5 @@
+import PrayerTitleList from "../../../../../src/domain/prayerTitles/components/prayerTitleList";
+
+export default function PrayerRoomBody(): JSX.Element {
+  return <PrayerTitleList />;
+}
