+++ babel.config.js
@@ -3,6 +3,16 @@ module.exports = function (api) {
   return {
     presets: ["babel-preset-expo"],
     plugins: [
+      // 경로 별칭 설정
+      [
+        "module-resolver",
+        {
+          root: ["./"],
+          alias: {
+            "@": "./src",
+          },
+        },
+      ],
       // NOTE: this is only necessary if you are using reanimated for animations
       "react-native-reanimated/plugin",
     ],
