+++ package.json
@@ -45,6 +45,7 @@
     "@babel/core": "^7.25.2",
     "@types/node": "^22.13.4",
     "@types/react": "~18.3.12",
+    "babel-plugin-module-resolver": "^5.0.2",
     "fs-extra": "^11.3.0",
     "typescript": "^5.7.3"
   },
