+++ src/domain/PrayerCreation/components/PrayerCreationBody.tsx
@@ -7,9 +7,9 @@ import { SelectedMember } from "../types/SelectedMember";
 import { KeyboardAwareScrollView } from "react-native-keyboard-aware-scroll-view";
 import { usePrayerCreationStore } from "../stores/usePrayerCreationStore";
 import { PrayerCreationItem } from "../types/PrayerCreationItem";
-import { useSelectedRoomStore } from "../../prayerRoom/stores/useSelectedRoomStore";
-import { useRoomMembersQuery } from "../../prayerRoom/hooks/queries/roomQueries";
-import { RoomMember } from "../../prayerRoom/types/dto/response/roomMember";
+import { useSelectedRoomStore } from "../../rooms/stores/useSelectedRoomStore";
+import { useRoomMembersQuery } from "../../rooms/hooks/queries/roomQueries";
+import { RoomMember } from "../../rooms/types/roomMember";
 
 // 컴포넌트 임포트
 import PrayerTitleInput from "./PrayerTitleInput";
