+++ src/domain/PrayerCreation/components/PrayerCreationBottom.tsx
@@ -7,7 +7,7 @@ import { useRouter } from "expo-router";
 import path from "../../../common/constants/path";
 import { usePrayerCreationMutation } from "../hooks/mutations/usePrayerCreationMutations";
 import { usePrayerCreationStore } from "../stores/usePrayerCreationStore";
-import { useSelectedRoomStore } from "../../prayerRoom/stores/useSelectedRoomStore";
+import { useSelectedRoomStore } from "../../rooms/stores/useSelectedRoomStore";
 import { useQueryClient } from "@tanstack/react-query";
 import QUERY_KEYS from "../../../common/hooks/queries/queryKeys";
