+++ src/domain/PrayerCreation/types/SelectedMember.ts
@@ -1,4 +1,4 @@
-import { RoomMember } from "../../prayerRoom/types/dto/response/roomMember";
+import { RoomMember } from "../../rooms/types/roomMember";
 
 export interface SelectedMember extends RoomMember {
   // isRoomMember: boolean;
