--- src/domain/prayerRoom/components/PrayerRoomBody.tsx
@@ -1,5 +1,0 @@
-import PrayerTitleList from "../../prayerTitles/components/prayerTitleList";
-
-export default function PrayerRoomBody(): JSX.Element {
-  return <PrayerTitleList />;
-}
