+++ src/domain/prayerUpdate/components/PrayerUpdateBody.tsx
@@ -5,9 +5,9 @@ import { RFValue } from "react-native-responsive-fontsize";
 import { color } from "../../../common/styles/color";
 import { SelectedMember } from "../../prayerCreation/types/SelectedMember";
 import { KeyboardAwareScrollView } from "react-native-keyboard-aware-scroll-view";
-import { useSelectedRoomStore } from "../../prayerRoom/stores/useSelectedRoomStore";
-import { useRoomMembersQuery } from "../../prayerRoom/hooks/queries/roomQueries";
-import { RoomMember } from "../../prayerRoom/types/dto/response/roomMember";
+import { useSelectedRoomStore } from "../../rooms/stores/useSelectedRoomStore";
+import { useRoomMembersQuery } from "../../rooms/hooks/queries/roomQueries";
+import { RoomMember } from "../../rooms/types/roomMember";
 import { PrayerUpdateItem } from "../types/prayerUpdateItem";
 import { usePrayerUpdateStore } from "../stores/usePrayerUpdateStore";
 import { useSelectedPrayerTitleStore } from "./../../prayerTitles/stores/useSelectedPrayerTitleStore";
