+++ src/domain/rooms/types/params/roomListParams.ts
@@ -1,0 +1,7 @@
+import { OrderBy, Dir } from "@/common/apis/constants/params";
+
+export interface RoomListParams {
+  orderBy: OrderBy;
+  after: string;
+  dir: Dir;
+}
