+++ src/domain/rooms/types/roomMember.ts
